fix:条件NOT判断 fix:RemainingAttemptsCondition条件解析和addProgress bug fix:TodayDepositCondition的match参数错误bug fix:ItemDropDataHolder重载bug fix:ConfigManager的notifyUpdateConfig bug fix:AbstractFriendRoomDao,AccountDao,SmsConfigDao,FriendRoom,CoreSendMessageManager NPE fix:AbstractRoomDao 锁时间获取异常,更新玩家座位 bug fix:CoreMessageHandler gm初始化玩家时初始化钻石,实际是金币的bug fix:ShopMessageHandler 兑换商城道具时发送给前端的数量错误的bug fix:AwardCodeManager 删除领奖码返回错误 fix:RedDotManager 初始化红点信息bug fix:SnowflakeManager 将主节点信息移动到cleanupExpiredNodes fix:SubscriptionManager 删除冗余代码 fix:AbstractPlayerService 操作保险箱,返回默认code修改 fix:PlayerSessionService remove方法时的反序列化修改 优化:RankService BATCH_ADD_LUA修改为point<=0时移除数据 fix:ShopService 清空商城时未更新内存数据 fix:任务红点bug fix:TaskService.java shouldDelete定时bug fix:PointsAwardLeaderboardManager 积分大奖排行榜并发bug

+++ core/src/main/java/com/jjg/game/core/manager/AwardCodeManager.java
@@ -81,6 +81,10 @@ public class AwardCodeManager {
      * @return 雪花ID
      */
     public long decode(String str) {
+        List<Long> decode = sqids.decode(str);
+        if (decode.isEmpty()) {
+            return 0;
+        }
         return sqids.decode(str).getFirst();
     }
 
@@ -151,8 +155,11 @@ public class AwardCodeManager {
      */
     public boolean deleteCode(String code) {
         long id = decode(code);
-        awardCodeDao.deleteById(id);
-        return true;
+        if (id > 0) {
+            awardCodeDao.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     /**
@@ -163,6 +170,9 @@ public class AwardCodeManager {
      */
     public boolean useCode(String code) {
         long id = decode(code);
+        if (id <= 0) {
+            return false;
+        }
         Optional<AwardCode> awardCodeOptional = awardCodeDao.findById(id);
         if (awardCodeOptional.isPresent()) {
             AwardCode awardCode = awardCodeOptional.get();
